% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aMiDAT.R
\name{aMiDAT}
\alias{aMiDAT}
\title{An adaptive microbiome-based direction-assisted test (aMiDAT) for microbiome differential
abundance analyses??}
\usage{
aMiDAT(X1, X2, lambdaSet, B)
}
\arguments{
\item{X1}{relative abundance observations of OTUs for the first sample; the rows represent individuals and
the columns represent OTUs.}

\item{X2}{relative abundance observations of OTUs for the second sample; the rows represent individuals and
the columns represent OTUs.}

\item{lambdaSet}{a vector containing candidate values for lambda.}

\item{B}{total number of permutations}
}
\value{
A list with the first argument being the aMiDAT statistic value and the second argument being
the p-value of aMiDAT based on permutations.
}
\description{
'aMiDAT()' returns the test statistic value of aMiDAT and its correpsonding p-value.
}
\details{
This function is used to calculate the adaptive microbiome-based direction-assisted test (aMiDAT)
statistic and its p-value for microbiome differential abundance analyses. aMiDAT combines multiple
MiDAT statistics by using their minimal p-value as the test statistic. The p-value of aMiDAT is
calcuated based on an one-layer permutation procedure.
}
\examples{
library(mvtnorm)
p = 20
n1 = 100
n2 = 100
mu1 = runif(p, min=0, max=10)
mu2 = mu1 + c(rep(0.5,5), rep(0,p-5))
sigMat1 = diag(p)
rho = 0.5
sigMat2 = rho^(abs(matrix(1:p, nrow=p, ncol=p, byrow=T)-matrix(1:p, nrow=p, ncol=p, byrow=F)))
Z1 = rmvnorm(n1, mean=mu1, sigma=sigMat1); W1 = exp(Z1)
Z2 = rmvnorm(n2, mean=mu2, sigma=sigMat2); W2 = exp(Z2)
X1 = W1/matrix(rowSums(W1), nrow=n1, ncol=p, byrow=F)
X2 = W2/matrix(rowSums(W2), nrow=n2, ncol=p, byrow=F)
aMiDAT(X1=X1, X2=X2, lambdaSet=c(1,2,3,4,Inf), B=1000)

}
