% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiDAT_stat.R
\name{MiDAT_stat}
\alias{MiDAT_stat}
\title{A microbiome-based direction-assisted test (MiDAT) for microbiome differential abundance analyses.}
\usage{
MiDAT_stat(X1, X2, lambdaSet)
}
\arguments{
\item{X1}{relative abundance observations of OTUs for the first sample; the rows represent individuals and
the columns represent OTUs.}

\item{X2}{relative abundance observations of OTUs for the second sample; the rows represent individuals and
the columns represent OTUs.}

\item{lambdaSet}{a vector containing candidate values for lambda.}
}
\value{
A vector containing the MiDAT statistics for all candidate values of lambda.
}
\description{
'MiDAT_stat()' returns the test statistic values of MiDAT for a set of cadidates lambda.
}
\details{
This function is used to calculate the microbiome-based direction-assisted test (MiDAT) statistics
for microbiome differential abundance analyses. MiDAT is developed to detect the overall
difference of micorbiome relative abundances between two or more biological/clinical conditions.
}
\examples{
p = 20
n1 = 100
n2 = 100
mu1 = runif(p, min=0, max=10)
mu2 = mu1 + c(rep(0.5,5), rep(0,p-5))
sigMat1 = diag(p)
rho = 0.5
sigMat2 = rho^(abs(matrix(1:p, nrow=p, ncol=p, byrow=TRUE)-matrix(1:p, nrow=p, ncol=p, byrow=FALSE)))
Z1 = MASS::mvrnorm(n1, mu=mu1, Sigma=sigMat1); W1 = exp(Z1)
Z2 = MASS::mvrnorm(n2, mu=mu2, Sigma=sigMat2); W2 = exp(Z2)
X1 = W1/matrix(rowSums(W1), nrow=n1, ncol=p, byrow=FALSE)
X2 = W2/matrix(rowSums(W2), nrow=n2, ncol=p, byrow=FALSE)
MiDAT_stat(X1=X1, X2=X2, lambdaSet=c(1,2,3,4,Inf))

}
